+++ src/test/java/org/apache/commons/lang3/mutable/MutableIntTest.java
@@ -150,7 +150,7 @@ public class MutableIntTest {
 
     @Test
     public void testIncrementAndGet() {
-        final MutableInt mutNum = new MutableInt((int) 1);
+        final MutableInt mutNum = new MutableInt(1);
         final int result = mutNum.incrementAndGet();
 
         assertEquals(2, result);
@@ -160,7 +160,7 @@ public class MutableIntTest {
 
     @Test
     public void testGetAndIncrement() {
-        final MutableInt mutNum = new MutableInt((int) 1);
+        final MutableInt mutNum = new MutableInt(1);
         final int result = mutNum.getAndIncrement();
 
         assertEquals(1, result);
@@ -179,7 +179,7 @@ public class MutableIntTest {
 
     @Test
     public void testDecrementAndGet() {
-        final MutableInt mutNum = new MutableInt((int) 1);
+        final MutableInt mutNum = new MutableInt(1);
         final int result = mutNum.decrementAndGet();
 
         assertEquals(0, result);
@@ -189,7 +189,7 @@ public class MutableIntTest {
 
     @Test
     public void testGetAndDecrement() {
-        final MutableInt mutNum = new MutableInt((int) 1);
+        final MutableInt mutNum = new MutableInt(1);
         final int result = mutNum.getAndDecrement();
 
         assertEquals(1, result);
@@ -217,38 +217,38 @@ public class MutableIntTest {
 
     @Test
     public void testGetAndAddValuePrimitive() {
-        final MutableInt mutableInteger = new MutableInt((int)0);
-        final int result = mutableInteger.getAndAdd((int) 1);
+        final MutableInt mutableInteger = new MutableInt(0);
+        final int result = mutableInteger.getAndAdd(1);
 
-        assertEquals((int) 0, result);
-        assertEquals((int) 1, mutableInteger.intValue());
+        assertEquals(0, result);
+        assertEquals(1, mutableInteger.intValue());
     }
 
     @Test
     public void testGetAndAddValueObject() {
-        final MutableInt mutableInteger = new MutableInt((int)0);
-        final int result = mutableInteger.getAndAdd(Integer.valueOf((int) 1));
+        final MutableInt mutableInteger = new MutableInt(0);
+        final int result = mutableInteger.getAndAdd(Integer.valueOf(1));
 
-        assertEquals((int) 0, result);
-        assertEquals((int) 1, mutableInteger.intValue());
+        assertEquals(0, result);
+        assertEquals(1, mutableInteger.intValue());
     }
 
     @Test
     public void testAddAndGetValuePrimitive() {
-        final MutableInt mutableInteger = new MutableInt((int)0);
-        final int result = mutableInteger.addAndGet((int) 1);
+        final MutableInt mutableInteger = new MutableInt(0);
+        final int result = mutableInteger.addAndGet(1);
 
-        assertEquals((int) 1, result);
-        assertEquals((int) 1, mutableInteger.intValue());
+        assertEquals(1, result);
+        assertEquals(1, mutableInteger.intValue());
     }
 
     @Test
     public void testAddAndGetValueObject() {
-        final MutableInt mutableInteger = new MutableInt((int)0);
-        final int result = mutableInteger.addAndGet(Integer.valueOf((int) 1));
+        final MutableInt mutableInteger = new MutableInt(0);
+        final int result = mutableInteger.addAndGet(Integer.valueOf(1));
 
-        assertEquals((int) 1, result);
-        assertEquals((int) 1, mutableInteger.intValue());
+        assertEquals(1, result);
+        assertEquals(1, mutableInteger.intValue());
     }
 
     @Test
